@@ -1,6 +1,6 @@
 {
   "name": "angular-cli",
-  "version": "5.1.0-alpha.20",
+  "version": "5.1.0-alpha.22",
   "private": true,
   "license": "MIT",
   "scripts": {
@@ -34,29 +34,29 @@
     "@angular-devkit/build-angular": "^0.13.4",
     "@angular/cli": "^7.3.6",
     "@angular/compiler-cli": "^7.2.6",
-    "@storybook/addon-actions": "5.1.0-alpha.20",
-    "@storybook/addon-backgrounds": "5.1.0-alpha.20",
-    "@storybook/addon-centered": "5.1.0-alpha.20",
-    "@storybook/addon-jest": "5.1.0-alpha.20",
-    "@storybook/addon-knobs": "5.1.0-alpha.20",
-    "@storybook/addon-links": "5.1.0-alpha.20",
-    "@storybook/addon-notes": "5.1.0-alpha.20",
-    "@storybook/addon-options": "5.1.0-alpha.20",
-    "@storybook/addon-storyshots": "5.1.0-alpha.20",
-    "@storybook/addon-storysource": "5.1.0-alpha.20",
-    "@storybook/addons": "5.1.0-alpha.20",
-    "@storybook/angular": "5.1.0-alpha.20",
+    "@storybook/addon-actions": "5.1.0-alpha.22",
+    "@storybook/addon-backgrounds": "5.1.0-alpha.22",
+    "@storybook/addon-centered": "5.1.0-alpha.22",
+    "@storybook/addon-jest": "5.1.0-alpha.22",
+    "@storybook/addon-knobs": "5.1.0-alpha.22",
+    "@storybook/addon-links": "5.1.0-alpha.22",
+    "@storybook/addon-notes": "5.1.0-alpha.22",
+    "@storybook/addon-options": "5.1.0-alpha.22",
+    "@storybook/addon-storyshots": "5.1.0-alpha.22",
+    "@storybook/addon-storysource": "5.1.0-alpha.22",
+    "@storybook/addons": "5.1.0-alpha.22",
+    "@storybook/angular": "5.1.0-alpha.22",
     "@types/core-js": "^2.5.0",
     "@types/jest": "^24.0.11",
     "@types/node": "~11.11.0",
     "@types/webpack-env": "^1.13.7",
     "global": "^4.3.2",
-    "jasmine-core": "~3.3.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^23.6.0",
-    "jest-preset-angular": "^6.0.1",
+    "jest": "^24.7.1",
+    "jest-preset-angular": "^7.0.1",
     "protractor": "~5.4.2",
     "ts-node": "~8.0.3",
-    "typescript": "^3.3.3333"
+    "typescript": "^3.4.1"
   }
 }
