@@ -1,6 +1,6 @@
 {
   "name": "cra-ts-kitchen-sink",
-  "version": "5.1.0-alpha.20",
+  "version": "5.1.0-alpha.22",
   "private": true,
   "scripts": {
     "build-storybook": "build-storybook -s public",
@@ -9,22 +9,16 @@
     "storybook": "start-storybook -p 9010 -s public",
     "test": "react-scripts test --env=jsdom"
   },
-  "jest": {
-    "snapshotSerializers": [
-      "jest-emotion",
-      "enzyme-to-json/serializer"
-    ]
-  },
   "dependencies": {
     "react": "^16.8.4",
     "react-dom": "^16.8.4"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "5.1.0-alpha.20",
-    "@storybook/addon-info": "5.1.0-alpha.20",
-    "@storybook/addon-options": "5.1.0-alpha.20",
-    "@storybook/addons": "5.1.0-alpha.20",
-    "@storybook/react": "5.1.0-alpha.20",
+    "@storybook/addon-actions": "5.1.0-alpha.22",
+    "@storybook/addon-info": "5.1.0-alpha.22",
+    "@storybook/addon-options": "5.1.0-alpha.22",
+    "@storybook/addons": "5.1.0-alpha.22",
+    "@storybook/react": "5.1.0-alpha.22",
     "@types/enzyme": "^3.9.0",
     "@types/react": "^16.8.4",
     "@types/react-dom": "^16.8.2",
@@ -36,9 +30,9 @@
     "enzyme-to-json": "^3.3.5",
     "fork-ts-checker-webpack-plugin": "^0.5.2",
     "react-docgen-typescript-loader": "^3.0.1",
-    "react-scripts": "^2.1.3",
+    "react-scripts": "^2.1.8",
     "tslint": "^5.14.0",
     "tslint-config-airbnb": "^5.11.1",
-    "typescript": "^3.3.3333"
+    "typescript": "^3.4.1"
   }
 }
