@@ -171,7 +181,7 @@
     "tslint": "~5.14.0",
     "tslint-config-prettier": "^1.18.0",
     "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "^3.3.3333",
+    "typescript": "^3.4.1",
     "weak": "^1.0.1"
   },
   "resolutions": {
