@@ -1,21 +1,34 @@
 module.exports = {
   presets: [
-    ['@babel/preset-env', { shippedProposals: true, useBuiltIns: 'usage' }],
+    ['@babel/preset-env', { shippedProposals: true, useBuiltIns: 'usage', corejs: '2' }],
+    '@babel/preset-typescript',
     '@babel/preset-react',
     '@babel/preset-flow',
   ],
   plugins: [
-    ['@babel/plugin-proposal-object-rest-spread', { loose: true, useBuiltIns: true }],
+    [
+      '@babel/plugin-proposal-decorators',
+      {
+        legacy: true,
+      },
+    ],
+    ['@babel/plugin-proposal-class-properties', { loose: true }],
     '@babel/plugin-proposal-export-default-from',
     '@babel/plugin-syntax-dynamic-import',
-    ['@babel/plugin-proposal-class-properties', { loose: true }],
+    ['@babel/plugin-proposal-object-rest-spread', { loose: true, useBuiltIns: true }],
     'babel-plugin-macros',
     ['emotion', { sourceMap: true, autoLabel: true }],
   ],
   env: {
     test: {
-      presets: [['@babel/preset-env', { shippedProposals: true, useBuiltIns: 'usage' }]],
-      plugins: ['babel-plugin-require-context-hook', 'babel-plugin-dynamic-import-node'],
+      presets: [
+        ['@babel/preset-env', { shippedProposals: true, useBuiltIns: 'usage', corejs: '2' }],
+      ],
+      plugins: [
+        'babel-plugin-require-context-hook',
+        'babel-plugin-dynamic-import-node',
+        '@babel/plugin-transform-runtime',
+      ],
     },
   },
   overrides: [
@@ -26,7 +39,7 @@ module.exports = {
     {
       test: './lib',
       presets: [
-        ['@babel/preset-env', { shippedProposals: true, useBuiltIns: 'usage' }],
+        ['@babel/preset-env', { shippedProposals: true, useBuiltIns: 'usage', corejs: '2' }],
         '@babel/preset-react',
       ],
       plugins: [
@@ -59,8 +72,16 @@ module.exports = {
             targets: {
               node: '8.11',
             },
+            corejs: '2',
           },
         ],
+      ],
+      plugins: [
+        'emotion',
+        'babel-plugin-macros',
+        ['@babel/plugin-proposal-class-properties', { loose: true }],
+        '@babel/plugin-proposal-object-rest-spread',
+        '@babel/plugin-proposal-export-default-from',
       ],
     },
   ],
