@@ -1,22 +1,43 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { styled, themes, convert } from '@storybook/theming';
 import memoize from 'memoizerific';
 
 import { NodeResult } from 'axe-core';
-import { Rules } from './Rules';
 import { RuleType } from '../A11YPanel';
 import { addElement } from '../../redux-config';
 import { IFRAME } from '../../constants';
-
-const Checkbox = styled.input({
-  cursor: 'pointer',
-});
 
 export class HighlightedElementData {
   originalOutline: string;
   isHighlighted: boolean;
 }
+
+interface ToggleProps {
+  elementsToHighlight: NodeResult[];
+  type: RuleType;
+  addElement?: (data: any) => void;
+  highlightedElementsMap?: Map<HTMLElement, HighlightedElementData>;
+  isToggledOn?: boolean;
+  toggleId?: string;
+  indeterminate?: boolean;
+}
+
+enum CheckBoxStates {
+  CHECKED,
+  UNCHECKED,
+  INDETERMINATE,
+}
+
+const Checkbox = styled.input(({ disabled }) => ({
+  cursor: disabled ? 'not-allowed' : 'pointer',
+}));
+
+const colorsByType = [
+  convert(themes.normal).color.negative, // VIOLATION,
+  convert(themes.normal).color.positive, // PASS,
+  convert(themes.normal).color.warning, // INCOMPLETION,
+];
 
 const getIframe = memoize(1)(() => document.getElementsByTagName(IFRAME)[0]);
 
@@ -28,106 +49,92 @@ function getElementBySelectorPath(elementPath: string): HTMLElement {
   return null;
 }
 
-function areAllRequiredElementsHiglighted(
-  elementsToHighlight: NodeResult[],
-  highlightedElementsMap: Map<HTMLElement, HighlightedElementData>
-): boolean {
-  let elementsInMapExist = false;
-  if (elementsToHighlight) {
-    for (let element of elementsToHighlight) {
-      if (element) {
-        const targetElement = getElementBySelectorPath(element.target[0]);
-        if (highlightedElementsMap.get(targetElement)) {
-          elementsInMapExist = true;
-          if (!highlightedElementsMap.get(targetElement).isHighlighted) {
-            return false;
-          }
-        }
-      }
-    }
-  }
-  return elementsInMapExist;
+function setElementOutlineStyle(targetElement: HTMLElement, outlineStyle: string): void {
+  targetElement.style.outline = outlineStyle;
 }
 
-interface ToggleProps {
-  elementsToHighlight: NodeResult[];
-  type: RuleType;
-  addElement?: (data: any) => void;
-  highlightedElementsMap?: Map<HTMLElement, HighlightedElementData>;
-  isToggledOn?: boolean;
-  toggleId?: string;
+function areAllRequiredElementsHighlighted(
+  elementsToHighlight: NodeResult[],
+  highlightedElementsMap: Map<HTMLElement, HighlightedElementData>
+): CheckBoxStates {
+  const highlightedCount = elementsToHighlight.filter(item => {
+    const targetElement = getElementBySelectorPath(item.target[0]);
+    return (
+      highlightedElementsMap.has(targetElement) &&
+      highlightedElementsMap.get(targetElement).isHighlighted
+    );
+  }).length;
+
+  return highlightedCount === 0
+    ? CheckBoxStates.UNCHECKED
+    : highlightedCount === elementsToHighlight.length
+    ? CheckBoxStates.CHECKED
+    : CheckBoxStates.INDETERMINATE;
 }
 
 function mapDispatchToProps(dispatch: any) {
   return {
-    addElement: (data: any) => dispatch(addElement(data)),
+    addElement: (data: { element: HTMLElement; data: HighlightedElementData }) =>
+      dispatch(addElement(data)),
   };
 }
 
 const mapStateToProps = (state: any, ownProps: any) => {
-  const isToggledOn = areAllRequiredElementsHiglighted(
-    ownProps.elementsToHighlight,
+  const checkBoxState = areAllRequiredElementsHighlighted(
+    ownProps.elementsToHighlight || [],
     state.highlightedElementsMap
   );
   return {
     highlightedElementsMap: state.highlightedElementsMap,
-    isToggledOn,
+    isToggledOn: checkBoxState === CheckBoxStates.CHECKED,
+    indeterminate: checkBoxState === CheckBoxStates.INDETERMINATE,
   };
 };
 
-class HighlightToggle extends Component<ToggleProps, {}> {
+class HighlightToggle extends Component<ToggleProps> {
   static defaultProps: Partial<ToggleProps> = {
     elementsToHighlight: [],
   };
 
+  private checkBoxRef = React.createRef<HTMLInputElement>();
+
   componentDidMount() {
-    for (let element of this.props.elementsToHighlight) {
-      if (element) {
-        const targetElement = getElementBySelectorPath(element.target[0]);
-        if (targetElement && !this.props.highlightedElementsMap.get(targetElement)) {
-          this.saveElementDataToMap(
-            targetElement,
-            false,
-            targetElement.style.outline,
-            this.props.type
-          );
-        }
+    this.props.elementsToHighlight.forEach(element => {
+      const targetElement = getElementBySelectorPath(element.target[0]);
+      if (targetElement && !this.props.highlightedElementsMap.has(targetElement)) {
+        this.saveElementDataToMap(targetElement, false, targetElement.style.outline);
       }
+    });
+  }
+
+  componentDidUpdate(prevProps: Readonly<ToggleProps>): void {
+    if (this.checkBoxRef.current) {
+      this.checkBoxRef.current.indeterminate = this.props.indeterminate;
     }
   }
 
-  higlightRuleLocation(targetElement: HTMLElement, addHighlight: boolean): void {
-    const OUTLINE_STYLE = `dotted`;
-    const OUTLINE_WIDTH = `1px`;
-    if (targetElement) {
-      if (addHighlight) {
-        switch (this.props.type) {
-          case RuleType.PASS:
-            this.setTargetElementOutlineStyle(targetElement, `${convert(themes.normal).color.positive} ${OUTLINE_STYLE} ${OUTLINE_WIDTH}`);
-            break;
-          case RuleType.VIOLATION:
-            this.setTargetElementOutlineStyle(targetElement, `${convert(themes.normal).color.negative} ${OUTLINE_STYLE} ${OUTLINE_WIDTH}`);
-            break;
-          case RuleType.INCOMPLETION:
-            this.setTargetElementOutlineStyle(targetElement, `${convert(themes.normal).color.warning} ${OUTLINE_STYLE} ${OUTLINE_WIDTH}`);
-            break;
-        }
-      } else {
-        if (this.props.highlightedElementsMap.get(targetElement)) {
-          this.setTargetElementOutlineStyle(
-            targetElement,
-            this.props.highlightedElementsMap.get(targetElement).originalOutline
-          );
-        }
-      }
+  highlightRuleLocation(targetElement: HTMLElement, addHighlight: boolean): void {
+    if (!targetElement) {
+      return;
+    }
+
+    if (addHighlight) {
+      setElementOutlineStyle(targetElement, `${colorsByType[this.props.type]} dotted 1px`);
+      return;
+    }
+
+    if (this.props.highlightedElementsMap.has(targetElement)) {
+      setElementOutlineStyle(
+        targetElement,
+        this.props.highlightedElementsMap.get(targetElement).originalOutline
+      );
     }
   }
 
   saveElementDataToMap(
     targetElement: HTMLElement,
     isHighlighted: boolean,
-    originalOutline: string,
-    ruleTypeState: RuleType
+    originalOutline: string
   ): void {
     const data: HighlightedElementData = new HighlightedElementData();
     data.isHighlighted = isHighlighted;
@@ -136,42 +143,32 @@ class HighlightToggle extends Component<ToggleProps, {}> {
     this.props.addElement(payload);
   }
 
-  setTargetElementOutlineStyle(targetElement: HTMLElement, outlineStyle: string): void {
-    targetElement.style.outline = outlineStyle;
-  }
-
-  onToggle(): void {
-    for (let element of this.props.elementsToHighlight) {
-      if (element) {
-        const targetElement = getElementBySelectorPath(element.target[0]);
-        if (this.props.highlightedElementsMap.get(targetElement)) {
-          let originalOutline = this.props.highlightedElementsMap.get(targetElement).originalOutline;
-          if (
-            this.props.isToggledOn &&
-            this.props.highlightedElementsMap.get(targetElement).isHighlighted
-          ) {
-            this.higlightRuleLocation(targetElement, false);
-            this.saveElementDataToMap(targetElement, false, originalOutline, this.props.type);
-          } else if (
-            !this.props.isToggledOn &&
-            !this.props.highlightedElementsMap.get(targetElement).isHighlighted
-          ) {
-            this.higlightRuleLocation(targetElement, true);
-            this.saveElementDataToMap(targetElement, true, originalOutline, this.props.type);
-          }
-        }
+  onToggle = (): void => {
+    this.props.elementsToHighlight.forEach(element => {
+      const targetElement = getElementBySelectorPath(element.target[0]);
+      if (!this.props.highlightedElementsMap.has(targetElement)) {
+        return;
       }
-    }
-  }
+      const originalOutline = this.props.highlightedElementsMap.get(targetElement).originalOutline;
+      const { isHighlighted } = this.props.highlightedElementsMap.get(targetElement);
+      const { isToggledOn } = this.props;
+      if ((isToggledOn && isHighlighted) || (!isToggledOn && !isHighlighted)) {
+        const addHighlight = !isToggledOn && !isHighlighted;
+        this.highlightRuleLocation(targetElement, addHighlight);
+        this.saveElementDataToMap(targetElement, addHighlight, originalOutline);
+      }
+    });
+  };
 
   render() {
     return (
       <Checkbox
+        ref={this.checkBoxRef}
         id={this.props.toggleId}
         type="checkbox"
         aria-label="Highlight result"
-        disabled={this.props.elementsToHighlight && this.props.elementsToHighlight.length === 0 ? true : false}
-        onChange={() => this.onToggle()}
+        disabled={!this.props.elementsToHighlight.length}
+        onChange={this.onToggle}
         checked={this.props.isToggledOn}
       />
     );
