@@ -12,11 +12,16 @@ export function action(name: string, options: ActionOptions = {}): HandlerFuncti
   const handler = function action(...args: any[]) {
     const channel = addons.getChannel();
     const id = uuid();
+    const minDepth = 5; // anything less is really just storybook internals
+
     const actionDisplayToEmit: ActionDisplay = {
       id,
       count: 0,
       data: { name, args },
-      options: actionOptions,
+      options: {
+        ...actionOptions,
+        depth: minDepth + (actionOptions.depth || 3),
+      },
     };
     channel.emit(EVENT_ID, actionDisplayToEmit);
   };
