@@ -461,7 +461,7 @@ const styles = `
   .markdown-body code::before,
   .markdown-body code::after {
     letter-spacing: -0.2em;
-    content: "\00a0";
+    content: " ";
   }
   .markdown-body pre {
     word-wrap: normal;
