@@ -1,28 +1,25 @@
 import React from 'react';
 import { View } from 'react-native';
 import PropTypes from 'prop-types';
-import Events from './constants';
+import Constants from './constants';
 
 export default class Container extends React.Component {
   constructor(props) {
     super(props);
     this.state = { background: props.initialBackground || '' };
-    this.onBackgroundChange = this.onBackgroundChange.bind(this);
   }
 
   componentDidMount() {
     const { channel } = this.props;
-    // Listen to the notes and render it.
-    channel.on(Events.UPDATE_BACKGROUND, this.onBackgroundChange);
+    channel.on(Constants.UPDATE_BACKGROUND, this.onBackgroundChange);
   }
 
-  // This is some cleanup tasks when the Notes panel is unmounting.
   componentWillUnmount() {
     const { channel } = this.props;
-    channel.removeListener(Events.UPDATE_BACKGROUND, this.onBackgroundChange);
+    channel.removeListener(Constants.UPDATE_BACKGROUND, this.onBackgroundChange);
   }
 
-  onBackgroundChange(background) {
+  onBackgroundChange = (background) => {
     this.setState({ background });
   }
 
