@@ -3,7 +3,7 @@ import addons, { makeDecorator } from '@storybook/addons';
 
 import EVENTS from './constants';
 
-function emitOptions(options) {
+function emitOptions(options: any) {
   const channel = addons.getChannel();
   if (!channel) {
     throw new Error(
@@ -21,7 +21,7 @@ function emitOptions(options) {
 // setOptions function will send Storybook UI options when the channel is
 // ready. If called before, options will be cached until it can be sent.
 let globalOptions = {};
-export const setOptions = deprecate(options => {
+export const setOptions = deprecate((options: any) => {
   globalOptions = options;
   emitOptions(options);
 }, '`setOptions(options)` is deprecated. Please use the `withOptions(options)` decorator globally.');
@@ -32,7 +32,7 @@ export const withOptions = makeDecorator({
   skipIfNoParametersOrOptions: false,
   wrapper: deprecate((getStory, context, { options: inputOptions, parameters }) => {
     // do not send hierachy related options over the channel
-    const { hierarchySeparator, hierarchyRootSeparator, ...change } = {
+    const { hierarchySeparator, hierarchyRootSeparator, ...change }: any = {
       ...globalOptions,
       ...inputOptions,
       ...parameters,
@@ -48,7 +48,7 @@ export const withOptions = makeDecorator({
 
     // MUTATION !
     // eslint-disable-next-line no-param-reassign
-    context.options = {
+    (context as any).options = {
       ...globalOptions,
       ...inputOptions,
       ...parameters,
@@ -61,7 +61,7 @@ export const withOptions = makeDecorator({
         ...inputOptions,
         ...parameters,
       },
-    });
+    } as any);
   }, 'withOptions is deprecated, use addParameters({ options: {} }) instead'),
 });
 
