@@ -13,6 +13,7 @@ import {
   EventEmitter,
   SimpleChanges,
   SimpleChange,
+  ChangeDetectorRef,
 } from '@angular/core';
 import { STORY } from '../app.token';
 import { NgStory, ICollection } from '../types';
@@ -31,6 +32,7 @@ export class AppComponent implements OnInit, OnDestroy {
 
   constructor(
     private cfr: ComponentFactoryResolver,
+    private changeDetectorRef: ChangeDetectorRef,
     @Inject(STORY) private data: Observable<NgStory>
   ) {}
 
@@ -38,12 +40,18 @@ export class AppComponent implements OnInit, OnDestroy {
     this.data.pipe(first()).subscribe((data: NgStory) => {
       this.target.clear();
       const compFactory = this.cfr.resolveComponentFactory(data.component);
-      const ref = this.target.createComponent(compFactory);
-      const instance = ref.instance;
+      const componentRef = this.target.createComponent(compFactory);
+      const instance = componentRef.instance;
+      // For some reason, manual change detection ref is only working when getting the ref from the injector (rather than componentRef.changeDetectorRef)
+      const childChangeDetectorRef: ChangeDetectorRef = componentRef.injector.get(
+        ChangeDetectorRef
+      );
 
       this.subscription = this.data.subscribe(newData => {
         this.setProps(instance, newData);
-        ref.changeDetectorRef.detectChanges();
+        childChangeDetectorRef.markForCheck();
+        // Must detect changes on the current component in order to update any changes in child component's @HostBinding properties (angular/angular#22560)
+        this.changeDetectorRef.detectChanges();
       });
     });
   }
