@@ -1,7 +1,7 @@
 module.exports = {
   coveragePathIgnorePatterns: ['/jest-config/', '/node_modules/'],
   preset: 'jest-preset-angular',
-  setupTestFrameworkScriptFile: './jest-config/setup.ts',
+  setupFilesAfterEnv: ['./jest-config/setup.ts'],
   snapshotSerializers: [
     '<rootDir>/../../node_modules/jest-preset-angular/AngularSnapshotSerializer.js',
     '<rootDir>/../../node_modules/jest-preset-angular/HTMLCommentSerializer.js',
