@@ -1,11 +1,9 @@
-import initStoryshots, { multiSnapshotWithOptions } from '@storybook/addon-storyshots';
 import path from 'path';
-import { render as renderer, configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+import { render as renderer } from 'enzyme';
 import { createSerializer as enzymeSerializer } from 'enzyme-to-json';
 import { createSerializer as emotionSerializer } from 'jest-emotion';
 
-configure({ adapter: new Adapter() });
+import initStoryshots, { multiSnapshotWithOptions } from '@storybook/addon-storyshots';
 
 initStoryshots({
   framework: 'react',
