@@ -26,10 +26,18 @@ module.exports = {
   ],
   transform: {
     '^.+\\.jsx?$': '<rootDir>/scripts/babel-jest.js',
-    '^.+\\.tsx?$': '<rootDir>/node_modules/ts-jest',
+    '^.+\\.tsx?$': '<rootDir>/scripts/babel-jest.js',
   },
-  testMatch: ['**/__tests__/**/*.(j|t)s?(x)', '**/?(*.)+(spek|test).(j|t)s?(x)'],
-  testPathIgnorePatterns: ['/node_modules/', '/dist/', 'addon-jest.test.js', '/cli/test/'],
+  testMatch: ['**/__tests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[jt]s?(x)'],
+  testPathIgnorePatterns: [
+    '/node_modules/',
+    '/dist/',
+    'addon-jest.test.js',
+    '/cli/test/',
+    '/examples/cra-kitchen-sink/src/App.test.js',
+    '/examples/cra-ts-kitchen-sink/src/components/',
+    '/examples/angular-cli/src/app/',
+  ],
   collectCoverage: false,
   collectCoverageFrom: [
     'app/**/*.{js,jsx,ts,tsx}',
@@ -46,9 +54,9 @@ module.exports = {
   ],
   snapshotSerializers: ['jest-emotion', 'enzyme-to-json/serializer'],
   coverageDirectory: 'coverage',
+  setupFilesAfterEnv: ['./scripts/jest.init.js'],
   coverageReporters: ['lcov'],
   testEnvironment: 'jest-environment-jsdom-thirteen',
-  setupTestFrameworkScriptFile: './scripts/jest.init.js',
   setupFiles: ['raf/polyfill'],
   testURL: 'http://localhost',
   modulePathIgnorePatterns: ['/dist/.*/__mocks__/'],
