@@ -48,15 +48,19 @@ export class PostmsgTransport {
    * the event will be stored in a buffer and sent when the window exists.
    * @param event
    */
-  send(event: ChannelEvent): Promise<any> {
+  send(event: ChannelEvent, options?: any): Promise<any> {
     const iframeWindow = this.getWindow();
     if (!iframeWindow) {
       return new Promise((resolve, reject) => {
         this.buffer.push({ event, resolve, reject });
       });
     }
+    let depth = 15;
+    if (options && Number.isInteger(options.depth)) {
+      depth = options.depth;
+    }
 
-    const data = stringify({ key: KEY, event }, { maxDepth: 15 });
+    const data = stringify({ key: KEY, event }, { maxDepth: depth });
 
     // TODO: investigate http://blog.teamtreehouse.com/cross-domain-messaging-with-postmessage
     // might replace '*' with document.location ?
