@@ -45,7 +45,8 @@
 // import 'classlist.js';  // Run `npm install --save classlist.js`.
 
 /** IE10 and IE11 requires the following for the Reflect API. */
-// import 'core-js/es6/reflect';
+import 'core-js/es6/reflect';
+import 'core-js/es7/reflect';
 
 /**
  * Web Animations `@angular/platform-browser/animations`
