@@ -77,9 +77,16 @@ export const getContext = (() => {
         addons.setChannel(channel);
       }
     }
-
-    const storyStore = new StoryStore({ channel });
-    const clientApi = new ClientApi({ storyStore, decorateStory });
+    let storyStore;
+    let clientApi;
+    if (typeof window !== 'undefined' && window.__STORYBOOK_CLIENT_API__) {
+      clientApi = window.__STORYBOOK_CLIENT_API__;
+      // eslint-disable-next-line no-underscore-dangle
+      storyStore = clientApi._storyStore;
+    } else {
+      storyStore = new StoryStore({ channel });
+      clientApi = new ClientApi({ storyStore, decorateStory });
+    }
     const { clearDecorators } = clientApi;
     const configApi = new ConfigApi({ clearDecorators, storyStore, channel, clientApi });
 
