@@ -1,4 +1,4 @@
-import { document } from 'global';
+import { document, window } from 'global';
 
 import start from './start';
 
@@ -6,6 +6,7 @@ jest.mock('@storybook/client-logger');
 jest.mock('global', () => ({
   navigator: { userAgent: 'browser', platform: '' },
   window: {
+    __STORYBOOK_CLIENT_API__: undefined,
     addEventListener: jest.fn(),
     location: { search: '' },
     history: { replaceState: jest.fn() },
@@ -18,6 +19,10 @@ jest.mock('global', () => ({
     location: { search: '?id=kind--story' },
   },
 }));
+
+afterEach(() => {
+  window.__STORYBOOK_CLIENT_API__ = undefined;
+});
 
 it('returns apis', () => {
   const render = jest.fn();
@@ -32,6 +37,21 @@ it('returns apis', () => {
       forceReRender: expect.any(Function),
     })
   );
+});
+
+it('reuses the current client api when the lib is reloaded', () => {
+  jest.useFakeTimers();
+  const render = jest.fn();
+
+  const { clientApi } = start(render);
+
+  const valueOfClientApi = window.__STORYBOOK_CLIENT_API__;
+
+  const { clientApi: newClientApi } = start(render);
+  jest.runAllTimers();
+
+  expect(clientApi).toEqual(newClientApi);
+  expect(clientApi).toEqual(valueOfClientApi);
 });
 
 it('calls render when you add a story', () => {
