@@ -23,7 +23,7 @@ async function compileManager(managerConfig, managerStartTime) {
         }
 
         if (stats && (stats.hasErrors() || stats.hasWarnings())) {
-          const { warnings, errors } = stats.toJson();
+          const { warnings, errors } = stats.toJson(managerConfig.stats);
 
           errors.forEach(e => logger.error(e));
           warnings.forEach(e => logger.error(e));
@@ -35,7 +35,7 @@ async function compileManager(managerConfig, managerStartTime) {
       }
 
       logger.trace({ message: '=> manager built', time: process.hrtime(managerStartTime) });
-      stats.toJson().warnings.forEach(e => logger.warn(e));
+      stats.toJson(managerConfig.stats).warnings.forEach(e => logger.warn(e));
 
       resolve(stats);
     });
@@ -52,8 +52,10 @@ async function watchPreview(previewConfig) {
       },
       (error, stats) => {
         if (!error) {
+          const statsConfig = previewConfig.stats ? previewConfig.stats : { colors: true };
+
           // eslint-disable-next-line no-console
-          console.log(stats.toString({ colors: true }));
+          console.log(stats.toString(statsConfig));
         } else {
           logger.error(error.message);
         }
@@ -77,7 +79,7 @@ async function compilePreview(previewConfig, previewStartTime) {
         }
 
         if (stats && (stats.hasErrors() || stats.hasWarnings())) {
-          const { warnings, errors } = stats.toJson();
+          const { warnings, errors } = stats.toJson(previewConfig.stats);
 
           errors.forEach(e => logger.error(e));
           warnings.forEach(e => logger.error(e));
@@ -86,7 +88,7 @@ async function compilePreview(previewConfig, previewStartTime) {
       }
 
       logger.trace({ message: '=> Preview built', time: process.hrtime(previewStartTime) });
-      stats.toJson().warnings.forEach(e => logger.warn(e));
+      stats.toJson(previewConfig.stats).warnings.forEach(e => logger.warn(e));
 
       return resolve(stats);
     });
