@@ -16,7 +16,10 @@ export default ({ configType }) => {
 
   return {
     presets: [
-      [require.resolve('@babel/preset-env'), { shippedProposals: true, useBuiltIns: 'usage', corejs: 2 }],
+      [
+        require.resolve('@babel/preset-env'),
+        { shippedProposals: true, useBuiltIns: 'usage', corejs: 2 },
+      ],
       ...prodPresets,
     ],
     plugins: [
