@@ -23,16 +23,17 @@ const run = () =>
           '@emotion/core',
           '@emotion/styled',
           '@storybook/addons',
+          '@storybook/api',
+          '@storybook/components',
           '@storybook/core-events',
           '@storybook/theming',
-          '@storybook/components',
           'airbnb-js-shims',
           'core-js/es6/symbol',
           'core-js/fn/array/iterator',
           'emotion-theming',
           'prop-types',
-          'react-dom',
           'react',
+          'react-dom',
           'regenerator-runtime/runtime',
           resolveLocal('../dist/index.js'),
         ],
